refactored type annotation to be able to resuse union types
