+++ tests/test_grid.py
@@ -4,7 +4,7 @@ import numpy as np
 
 import tidy3d as td
 from tidy3d.components.grid import Coords, FieldGrid, YeeGrid, Grid
-from tidy3d.components.base import TYPE_TAG_STR
+from tidy3d.components.types import TYPE_TAG_STR
 
 
 def test_coords():
