+++ tidy3d/components/source.py
@@ -4,11 +4,11 @@ from abc import ABC, abstractmethod
 from typing import Union, Tuple
 import logging
 
-from typing_extensions import Annotated, Literal
+from typing_extensions import Literal
 import pydantic
 import numpy as np
 
-from .base import Tidy3dBaseModel, cached_property, TYPE_TAG_STR
+from .base import Tidy3dBaseModel, cached_property
 from .types import Direction, Polarization, Ax, FreqBound, ArrayLike, Axis, Bound
 from .validators import assert_plane, validate_name_str
 from .geometry import Box
@@ -673,14 +673,11 @@ class TFSF(AngledFieldSource, VolumeSource):
 
 
 # sources allowed in Simulation.sources
-SourceType = Annotated[
-    Union[
-        UniformCurrentSource,
-        PointDipole,
-        GaussianBeam,
-        AstigmaticGaussianBeam,
-        ModeSource,
-        PlaneWave,
-    ],
-    pydantic.Field(discriminator=TYPE_TAG_STR),
+SourceType = Union[
+    UniformCurrentSource,
+    PointDipole,
+    GaussianBeam,
+    AstigmaticGaussianBeam,
+    ModeSource,
+    PlaneWave,
 ]
